+++ setup-squirrel-windows-s3.js
